Move ApplicationModel into a separate file
